+++ tests/win32_error.c
@@ -4,6 +4,6 @@
 int main(void) {
     nob_log(NOB_ERROR, "First 100 Win32 API Errors:");
     for (DWORD i = 0; i < 100; i++)
-        nob_log(NOB_ERROR, "%lu (0x%X): \"%s\"", i, i, nob_win32_error_message(i));
+        nob_log(NOB_ERROR, "%u (0x%X): \"%s\"", (unsigned int)i, (unsigned int)i, nob_win32_error_message(i));
     return 0;
 }
